requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.8
# by the following command:
#
#    pip-compile
#
affine==2.4.0
    # via
    #   rasterio
    #   rasterstats
antlr4-python3-runtime==4.9.3
    # via
    #   hydra-core
    #   omegaconf
argparse==1.4.0
    # via -r requirements.in
attrs==23.2.0
    # via
    #   fiona
    #   rasterio
certifi==2024.2.2
    # via
    #   fiona
    #   pyproj
    #   rasterio
click==8.1.7
    # via
    #   click-plugins
    #   cligj
    #   fiona
    #   rasterio
    #   rasterstats
click-plugins==1.1.1
    # via
    #   fiona
    #   rasterio
cligj==0.7.2
    # via
    #   fiona
    #   rasterio
    #   rasterstats
colorama==0.4.6
    # via
    #   click
    #   loguru
    #   tqdm
csv-file==1.0
    # via -r requirements.in
docutils==0.20.1
    # via statistics
fiona==1.9.5
    # via
    #   geopandas
    #   rasterstats
geopandas==0.13.2
    # via -r requirements.in
glob2==0.7
    # via -r requirements.in
hydra-core==1.3.2
    # via -r requirements.in
importlib-metadata==7.0.1
    # via fiona
importlib-resources==6.1.2
    # via hydra-core
joblib==1.3.2
    # via
    #   -r requirements.in
    #   scikit-learn
loguru==0.7.2
    # via -r requirements.in
numpy==1.24.4
    # via
    #   -r requirements.in
    #   pandas
    #   rasterio
    #   rasterstats
    #   scikit-learn
    #   scipy
    #   shapely
    #   snuggs
omegaconf==2.3.0
    # via
    #   -r requirements.in
    #   hydra-core
packaging==23.2
    # via
    #   geopandas
    #   hydra-core
pandas==2.0.3
    # via geopandas
pyparsing==3.1.1
    # via snuggs
pyproj==3.5.0
    # via geopandas
python-dateutil==2.9.0.post0
    # via pandas
pytz==2024.1
    # via pandas
pyyaml==6.0.1
    # via
    #   -r requirements.in
    #   omegaconf
rasterio==1.3.9
    # via
    #   -r requirements.in
    #   rasterstats
rasterstats==0.19.0
    # via -r requirements.in
scikit-learn==1.3.2
    # via -r requirements.in
scipy==1.10.1
    # via scikit-learn
shapely==2.0.3
    # via
    #   geopandas
    #   rasterstats
simplejson==3.19.2
    # via rasterstats
six==1.16.0
    # via
    #   fiona
    #   python-dateutil
snuggs==1.4.7
    # via rasterio
statistics==1.0.3.5
    # via -r requirements.in
threadpoolctl==3.3.0
    # via scikit-learn
tqdm==4.66.2
    # via -r requirements.in
tzdata==2024.1
    # via pandas
win32-setctime==1.1.0
    # via loguru
zipp==3.17.0
    # via
    #   importlib-metadata
    #   importlib-resources

# The following packages are considered to be unsafe in a requirements file:
# setuptools
